Add files via upload
In this react project, I created a form componant  that stores data in the local Storage of the browser once they are submitted, the stored data will later be displayed in  a table componant. Storing data  in the local Storage is just a temporary solution for demonstration purposes, since JS has no direct access to the computer's local ressources.